refactor: extract batch planning
